### 2020-05-22 #2226 @whedon
[PRE REVIEW]: TEfits: Nonlinear regression for time-evolving indices
**Submitting author:** @akcochrane (<a href="http://orcid.org/0000-0001-6691-9149">Aaron Cochrane</a>)
**Repository:** <a href="https://github.com/akcochrane/TEfits" target ="_blank">https://github.com/akcochrane/TEfits</a>
**Version:** v0.77.01
