+++ R/TEhelp.R
@@ -8,14 +8,14 @@
 #'
 #' @export
 #' @examples
-#' ## Package summary:
+#' ## Brief package summary:
 #' TEhelp('package')
+#'
+#' ## List topics with TEhelp entries:
+#' TEhelp('topics')
 #'
 #' ## Overview of error functions:
 #' TEhelp('errFun')
-#'
-#' ## The d prime link function:
-#' TEhelp('linkFun=d_prime')
 #'
 #' ## The log-cosh error function:
 #' TEhelp('errFun=logcosh')
@@ -23,89 +23,146 @@
 #' ## The ex-Gaussian error function, with change in the exponential component:
 #' TEhelp('errFun=exGauss_tau')
 #'
+#' ## The d prime link function:
+#' TEhelp('linkFun=d_prime')
+#'
 #' ## The summary method for TEfit models:
 #' TEhelp('summary(TEfit)')
 #'
 TEhelp <- function(topic='package'){
-  if(topic=='package'){cat(
+
+  helpTable <- data.frame(matrix(NA,0,2))
+  colnames(helpTable) <- c('topic','description')
+
+  helpTable <-rbind(helpTable,data.frame(topic='package',description=paste0(
     'This package is designed to facilitate painless nonlinear regression of time-evolving measures.
 Every effort is made to balance sensible defaults, full customization, and interpretable parameterizations.
 
-    Use this help function to get more information about specific aspects. See ?TEhelp.'
-  )}
+For more information see ?TEfits, the README at github.com/akcochrane/TEfits, and/or vignette("TEfit_tutorial")
 
-  if(topic=='errFun'){cat(
-    "Various functions are available to optimize."
-    ,'These include:\nSum of squared error (`ols`; the default),'
-    ,'\nRoot mean squared error (`rmse`),'
-    ,'\nBernoulli (binary binomial; `bernoulli`),'
-    ,'\nlog-hyperbolic-cosine (`logcosh`).
+Use this help function to get more information about specific aspects. See ?TEhelp.'
+  )))
+
+
+  ## ## ##
+  ## error functions
+  {
+    helpTable <-rbind(helpTable,data.frame(topic='errFun',description=paste0(
+      "Various functions are available to optimize."
+      ,'These include:\nSum of squared error (`ols`; the default),'
+      ,'\nRoot mean squared error (`rmse`),'
+      ,'\nBernoulli (binary binomial; `bernoulli`),'
+      ,'\nlog-hyperbolic-cosine (`logcosh`).
 
     These are specified as an argument to TEfit(..., errFun="exGauss_tau", ...), TEfitAll(..., errFun="logcosh", ...), etc.
     ___ ___ ___ ___'
-  )}
+    )))
 
-  if(topic=='linkFun=logistic'){cat('help page pending')}
+    helpTable <-rbind(helpTable,data.frame(topic='errFun=logcosh',description=paste0(
+      'The log-hyperbolic-cosine error function'
+      ,'is, conceptually, a compromise between OLS and absolute error'
+      ,'\nfunctions. While at absolute errors of approximately'
+      ,'4 or greater the function is functionally'
+      ,"\nindistinguishable from absolute error, at smaller errors the function's"
+      ,'slope decreases,'
+      ,'\nwhich allows for easier optimization. In practice,'
+      ,'this is a versatile error function that'
+      ,'\nis especially well-suited for data with outliers or other shapes'
+      ,'indicating\ndispersion beyond a standard Gaussian distribution.'
+      ,'\n\nExample: mod <- TEfit(dat[,c("resp","trialNum")],errFun="logCosh")'
+      ,'\n\nR Implementation: log(cosh(y-yHat))\n__ __ __ __'
+    )))
 
[…]
