### 2020-05-23 3e234f0
Update TEhelp.R
