+++ man/TEglm.Rd
@@ -33,8 +33,7 @@ The value of rate has a lower bound of
 the .0333 quantile of the time variable (i.e., 87.5\% of change happens in the first 10\% of time) and an upper bound of the
 .333 quantile of the time variable (i.e., 87.5\% of change takes 100\% of the time to happen). These bounds provide
 some robustness in estimates of asympototic effects (i.e., "controlling for time") as well as initial effects
-(i.e., "time-related starting offset"). Uses this transformed time variable in a \code{\link{tef_rlm_boot}} model to estimate
-bootstrapped parameter coefficients and out-of-sample prediction.
+(i.e., "time-related starting offset").
 
 Mean estimated rate is calculated after trimming the upper 25\% and lower 25\% of bootstrapped rate estimates, for robustness to
 extremes in resampling.
