### 2020-06-11 0da4578
clarify parameterization of rates
