### 2020-06-11 d541883
Minor updates to optimization and diagnostics
1. Changed default optim method to L-BFGS-B
2. Change "p-value of Spearman change" to only indicate "rejection of the null" when the model correlation is lower than the null correlation
3. Minor documentation updates
