### 2020-06-27 #2226 @cMadan
@akcochrane, sorry for the delay, it looks like you need to add some `{` and `}` around terms in your bib to fix the capitalisation in the references but otherwise things look good.

Do you have any additional suggestions for reviewers (see the initial comment in this issue for some guidance) ?
