### 2020-06-29 ed0e0ea
Minor updates to documentation
