+++ man/tef_fitAll2brms.Rd
@@ -33,7 +33,8 @@ A \code{\link[brms]{brms}} nonlinear mixed-effects model object.
 Passes a \code{\link{TEfitAll}} model to [nonlinear mixed-effects Bayesian] fitting using the
 \code{\link[brms]{brms}} package. Note that, due to the extensive time needed to
 fit \code{\link[brms]{brms}} models,
-this function is less tested than most functions in the \code{TEfits} package.
+this function is less tested than most functions in the \code{TEfits} package. Functionality is
+\strong{experimental}.
 }
 \details{
 \code{TEfitAll} \code{bernoulli} models are fit using either \code{bernoulli} or \code{Beta} response
