+++ inst/references.bib
@@ -13,7 +13,7 @@
 }
 
 @article{burkner_brms_2017,
-	title = {brms: An R Package for Bayesian Multilevel Models Using Stan},
+	title = {brms: An {R} Package for Bayesian Multilevel Models Using Stan},
 	volume = {80},
 	issn = {1548-7660},
 	shorttitle = {\textbf{brms}},
@@ -44,7 +44,7 @@
 }
 
 @Book{venables_ripley,
-    title = {Modern Applied Statistics with S},
+    title = {Modern Applied Statistics with {S}},
     author = {W. N. Venables and B. D. Ripley},
     publisher = {Springer},
     edition = {Fourth},
