+++ man/TEfit.Rd
@@ -137,6 +137,11 @@ This is enforced by penalizing the time-evolving model's error multiplicatively 
 between the average of the model prediction and the average of the null [non-time-evolving] prediction. This is intended to
 constrain model predictions to a "sane" range. This constraint can be removed with \code{control=tef_control(penalizeMean=F)}.
 For this and other default model constraints see \code{?\link{tef_getBounds}}.
+
+Currently \strong{known bugs} are:
+\itemize{
+\item{The logistic linkfun must include threshChange=T. Returns error if only the bias term is allowed to change.}
+}
 }
 \examples{
 \dontrun{
