+++ doc/references.bib
@@ -1,0 +1,55 @@
+@article{bates_fitting_2015,
+	title = {Fitting Linear Mixed-Effects Models Using lme4},
+	volume = {67},
+	issn = {1548-7660},
+	url = {http://www.jstatsoft.org/v67/i01/},
+	doi = {10.18637/jss.v067.i01},
+	language = {en},
+	number = {1},
+	urldate = {2018-05-18},
+	journal = {Journal of Statistical Software},
+	author = {Bates, Douglas and Machler, Martin and Bolker, Ben and Walker, Steve},
+	year = {2015}
+}
+
+@article{burkner_brms_2017,
+	title = {brms: An {R} Package for Bayesian Multilevel Models Using Stan},
+	volume = {80},
+	issn = {1548-7660},
+	shorttitle = {\textbf{brms}},
+	url = {http://www.jstatsoft.org/v80/i01/},
+	doi = {10.18637/jss.v080.i01},
+	language = {en},
+	number = {1},
+	urldate = {2018-11-05},
+	journal = {Journal of Statistical Software},
+	author = {Buerkner, Paul-Christian},
+	year = {2017}
+}
+
+@article{kattner_trial-dependent_2017,
+	title = {Trial-dependent psychometric functions accounting for perceptual learning in 2-{AFC} discrimination tasks},
+	volume = {17},
+	issn = {1534-7362},
+	url = {http://jov.arvojournals.org/article.aspx?doi=10.1167/17.11.3},
+	doi = {10.1167/17.11.3},
+	language = {en},
+	number = {11},
+	urldate = {2017-10-22},
+	journal = {Journal of Vision},
+	author = {Kattner, Florian and Cochrane, Aaron and Green, C. Shawn},
+	month = sep,
+	year = {2017},
+	pages = {3}
+}
+
+@Book{venables_ripley,
+    title = {Modern Applied Statistics with {S}},
+    author = {W. N. Venables and B. D. Ripley},
+    publisher = {Springer},
+    edition = {Fourth},
+    address = {New York},
+    year = {2002},
+    note = {ISBN 0-387-95457-0},
+    url = {http://www.stats.ox.ac.uk/pub/MASS4},
+  }
