+++ doc/paper.md
@@ -14,7 +14,7 @@ affiliations:
 - index: 1
   name: University of Wisconsin - Madison
 output: html_document
-date: 1 May 2020
+date: 29 June 2020
 bibliography: references.bib
 ---
 
