+++ doc/paper.md
@@ -1,4 +1,4 @@
-
+---
 title: 'TEfits: Nonlinear regression for time-evolving indices'
 authors:
 - affiliation: 1
