+++ doc/references.bib
@@ -13,7 +13,7 @@
 }
 
 @article{burkner_brms_2017,
-	title = {brms: An {R} Package for Bayesian Multilevel Models Using Stan},
+	title = {brms: An {R} Package for Bayesian Multilevel Models Using {S}tan},
 	volume = {80},
 	issn = {1548-7660},
 	shorttitle = {\textbf{brms}},
