+++ doc/references.bib
@@ -56,8 +56,7 @@
 
 
 @article{johnston_perceptual,
-	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
-	year = {in press},
+	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images (in press)},
 	journal = {Journal of Vision},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
