+++ doc/references.bib
@@ -56,7 +56,8 @@
 
 
 @article{johnston_perceptual,
-	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images (in press)},
+	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
 	journal = {Journal of Vision},
+	year = {in press},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
