+++ doc/references.bib
@@ -53,3 +53,10 @@
     note = {ISBN 0-387-95457-0},
     url = {http://www.stats.ox.ac.uk/pub/MASS4},
   }
+  
+  @article{johnston_perceptual,
+	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
+	journal = {Journal of Vision},
+	year = "in press",
+	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
+}
