+++ doc/references.bib
@@ -57,7 +57,7 @@
 
 @article{johnston_perceptual,
 	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
-	journal = {Journal of Vision},
-	year = {"in press"},
+	journal = {Journal of Vision. (In press)},
+	year = {2020},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
