+++ doc/references.bib
@@ -1,3 +1,5 @@
+@Preamble{ " \newcommand{\noop}[1]{} " }
+
 @article{bates_fitting_2015,
 	title = {Fitting Linear Mixed-Effects Models Using lme4},
 	volume = {67},
@@ -56,8 +58,8 @@
 
 @article{johnston_perceptual,
 	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
-	journal = {Journal of Vision. (In press)},
-	year = {2020},
+	journal = {Journal of Vision},
+	year = "\noop{2020}in press",
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
 
