+++ doc/references.bib
@@ -59,7 +59,7 @@
 @article{johnston_perceptual,
 	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
 	journal = {Journal of Vision},
-	year = "\noop{2020}in press",
+	year = {"\noop{2020}in press"},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
 
