+++ doc/references.bib
@@ -56,10 +56,10 @@
     url = {http://www.stats.ox.ac.uk/pub/MASS4},
   }
 
-@article{johnston_perceptual,
+@Article{johnston_perceptual,
 	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
 	journal = {Journal of Vision},
-	year = {"\noop{2020}in press"},
+	year = {"in press"},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
 
