+++ doc/references.bib
@@ -59,7 +59,8 @@
 @Article{johnston_perceptual,
 	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
 	journal = {Journal of Vision},
-	year = {"in press"},
+	year = {2020},
+	volume = {"in press"},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
 
