+++ doc/references.bib
@@ -60,7 +60,7 @@
 	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
 	journal = {Journal of Vision},
 	year = {2020},
-	volume = {"in press"},
+	volume = {in press},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
 
