### 2020-06-29 #2226 @akcochrane
@cMadan Thanks for pointing those out, I've made some minor changes to the paper that include the capitalization of references.
Regarding reviewers, I don't have any particularly well-informed suggestions. Given the information available in the initial comment, it seems that **samhforbes**, **ssriv**, or **Darthpathos** may have relevant overlapping interest and expertise.
