### 2020-07-03 eb569f2
fixed bug with interface between TEglmem and TEglm
