### 2020-07-04 4ab14a9
Minor fix and addition
Fixed interface between TElmem and TElm.
Added initial functionality for smoothing data according to a Drift Diffusion Model.
