+++ NAMESPACE
@@ -43,3 +43,4 @@ export(tef_runningMean)
 export(tef_subsetCorrel)
 export(tef_tryFits)
 export(tef_vars2forms)
+export(tef_windowWDM)
