### 2020-07-04 81b0fa0
Update tef_windowWDM.R
Make tef_windowWDM  check for NAs
