### 2020-07-10 bf4cb42
Fixed bug in bootstrapped coefficient names
