+++ R/coef.TEfit.R
@@ -14,8 +14,8 @@ coef.TEfit <- function(TEs3){
   if(exists('bootList',TEs3)){
     nPars <- ncol(TEs3$bootList$boots)-4
 
-    bootCoef <- data.frame(Q025 = round(TEs3$bootList$boots[2,1:nPars],3),
-    Q975 =round(TEs3$bootList$boots[10,1:nPars],3))
+    bootCoef <- data.frame(Q025 = round(TEs3$bootList$boots['2.5%',1:nPars],3),
+    Q975 =round(TEs3$bootList$boots['97.5%',1:nPars],3))
 
     parVals <- merge(parVals,bootCoef,all=T,by=0)
 
