### 2020-07-10 bd397fd
Add example to TEfitAll
