### 2020-07-10 0a76c1d
Update TEfitAll.R
