### 2020-07-10 d169eb6
Update TEfitAll.Rd
