+++ man/TEfitAll.Rd
@@ -13,7 +13,7 @@ TEfitAll(
   linkFun = list(link = "identity"),
   errFun = "ols",
   changeFun = "expo",
-  bootPars = list(nBoots = 0, bootTries = 0, bootPercent = 0),
+  bootPars = tef_bootList(),
   blockTimeVar = NULL,
   covarTerms = list(),
   control = tef_control()
@@ -50,6 +50,13 @@ for every unique value of groupingVar. Defaults to
 returning a list including two summaries and all models; returning
 only a summary is also an option. Arguments (except groupingvar, a grouping vector)
 are identical to, and are passed directly to, \code{\link{TEfit}}.
+}
+\examples{
+\dontrun{
+m <- TEfitAll(anstrain[,c('acc','trialNum')],groupingVar = anstrain$subID,groupingVarName = 'subID',bootPars = tef_bootList(resamples = 20))
+summary(m)
+}
+
 }
 \seealso{
 \code{\link{TEfit}} for fitting a single model;
