### 2020-07-14 bcf1da1
Improve robustness to poor start values
