+++ man/TEfitAll.Rd
@@ -60,5 +60,5 @@ summary(m)
 }
 \seealso{
 \code{\link{TEfit}} for fitting a single model;
-\code{\link{tef_fitAll2brms}} to re-fit the TEfitAll output using \code{\link[brms]{brms}}
+\code{\link{tef_fitAll2brms}} to re-fit the TEfitAll output using \code{\link[brms]{brms-package}}
 }
