+++ man/tef_modestMouse.Rd
@@ -1,11 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tef_modestMouse.R
-\name{tef_modestMouse}
-\alias{tef_modestMouse}
-\title{Print a line from a Modest Mouse song.}
-\usage{
-tef_modestMouse()
-}
-\description{
-Mostly from This Is a Long Drive for Someone with Nothing to Think About
-}
