+++ NAMESPACE
@@ -36,7 +36,6 @@ export(tef_getTermsExpoFatigue)
 export(tef_getTermsPower4)
 export(tef_getTermsWeibull)
 export(tef_makeErrTable)
-export(tef_modestMouse)
 export(tef_penalizedErr)
 export(tef_rlm_boot)
 export(tef_runningMean)
