### 2020-07-14 693b4f8
Update .travis.yml
Attempting to avoid Travis-CI error "installation of package ‘V8’ had non-zero exit status"
