### 2020-07-14 032e217
Update .travis.yml
Attempting to avoid Travis-CI error "installation of package ‘V8’ had non-zero exit status"
