+++ .travis.yml
@@ -3,6 +3,8 @@ language: r
 warnings_are_errors: false
 
 r_packages:
+  - remotes
+  - V8
   - MASS
   - testthat
 
