### 2020-07-14 17d0bb3
remove brms suggestion
Attempting to avoid Travis-CI error "installation of package ‘V8’ had non-zero exit status"
