+++ DESCRIPTION
@@ -9,6 +9,6 @@ License: MIT
 Encoding: UTF-8
 LazyData: true
 Imports: graphics, MASS, stats
-Suggests: knitr, rmarkdown, brms, psych, lme4
+Suggests: knitr, rmarkdown, psych, lme4
 VignetteBuilder: knitr
 RoxygenNote: 7.1.0
