### 2020-07-17 31bdbac
include robustness to errors when checking for convergence
