### 2020-07-17 bd44559
prevent infinite boundaries
