### 2020-07-18 a2cb362
Add tests for TEfitAll
