+++ R/tef_bootFits.R
@@ -109,7 +109,6 @@ tef_bootFits <- function(modList){
                          c(.01,.025,.1,pnorm(-1),.25,.5,.75,pnorm(1),.9,.975,.99),na.rm=T)
   bootList$bootCorrel <- cor(bootList$bootFits[,1:(dim(bootList$bootFits)[2]-1)])
 
-    bootList$percent_increasing <- mean((ending_predicted_values-beginning_predicted_values)>0)
 
   ## find the trial at which predicted values diverge from initial values with d>=1
   try({
@@ -135,6 +134,9 @@ tef_bootFits <- function(modList){
 
   ending_predicted_values <- bootList$bootPreds[,dim(bootList$bootPreds)[2]]
   beginning_predicted_values <- bootList$bootPreds[,1]
+  bootList$percent_increasing <- mean((ending_predicted_values-beginning_predicted_values)>0)
+
+
    bootList$slope_nonzero_D <-
     (mean(ending_predicted_values) - mean(beginning_predicted_values))/
     mean(c(sd(ending_predicted_values),sd(beginning_predicted_values)))
