### 2020-07-31 #2535 @whedon
[REVIEW]: TEfits: Nonlinear regression for time-evolving indices
**Submitting author:** @akcochrane (<a href="http://orcid.org/0000-0001-6691-9149">Aaron Cochrane</a>)
**Repository:** <a href="https://github.com/akcochrane/TEfits" target ="_blank">https://github.com/akcochrane/TEfits</a>
**Version:** v0.77.01
