+++ R/TEhelp.R
@@ -1,4 +1,5 @@
 
+
 #' Get help for the TEfits package
 #'
 #' Many options for fitting using TEfits are specified in a variety of places. To help,
@@ -30,141 +31,231 @@
 #' ## The summary method for TEfit models:
 #' TEhelp('summary(TEfit)')
 #'
-TEhelp <- function(topic='package'){
+TEhelp <- function(topic = 'package') {
+  helpTable <- data.frame(matrix(NA, 0, 2))
+  colnames(helpTable) <- c('topic', 'description')
 
-  helpTable <- data.frame(matrix(NA,0,2))
-  colnames(helpTable) <- c('topic','description')
+  helpTable <-
+    rbind(
+      helpTable,
+      data.frame(
+        topic = 'package',
+        description = paste0(
+          'This package is designed to facilitate painless nonlinear regression of time-evolving measures.
+          Every effort is made to balance sensible defaults, full customization, and interpretable parameterizations.
 
-  helpTable <-rbind(helpTable,data.frame(topic='package',description=paste0(
-    'This package is designed to facilitate painless nonlinear regression of time-evolving measures.
-Every effort is made to balance sensible defaults, full customization, and interpretable parameterizations.
+          For more information see ?TEfits, the README at github.com/akcochrane/TEfits, and/or vignette("TEfit_tutorial").
 
-For more information see ?TEfits, the README at github.com/akcochrane/TEfits, and/or vignette("TEfit_tutorial")
-
-Use this help function to get more information about specific aspects. See ?TEhelp.'
-  ),stringsAsFactors = TRUE))
+          Use this help function to get more information about specific aspects. See ?TEhelp.'
+        ),
+        stringsAsFactors = TRUE
+        )
+      )
 
 
   ## ## ##
   ## error functions
   {
-    helpTable <-rbind(helpTable,data.frame(topic='errFun',description=paste0(
-      "Various functions are available to optimize."
-      ,'These include:\nSum of squared error (`ols`; the default),'
-      ,'\nRoot mean squared error (`rmse`),'
-      ,'\nBernoulli (binary binomial; `bernoulli`),'
-      ,'\nlog-hyperbolic-cosine (`logcosh`).
+    helpTable <-
+      rbind(helpTable,
+            data.frame(
+              topic = 'errFun',
+              description = paste0(
+                "Various functions are available to optimize."
+                ,
+                'These include:\nSum of squared error (`ols`; the default),'
+                ,
+                '\nRoot mean squared error (`rmse`),'
+                ,
+                '\nBernoulli (binary binomial; `bernoulli`),'
+                ,
+                '\nlog-hyperbolic-cosine (`logcosh`).
 
-    These are specified as an argument to TEfit(..., errFun="exGauss_tau", ...), TEfitAll(..., errFun="logcosh", ...), etc.
-    ___ ___ ___ ___'
-    )))
+                These are specified as an argument to TEfit(..., errFun="exGauss_tau", ...), TEfitAll(..., errFun="logcosh", ...), etc.
+                ___ ___ ___ ___'
+              )
+              ))
 
-    helpTable <-rbind(helpTable,data.frame(topic='errFun=logcosh',description=paste0(
-      'The log-hyperbolic-cosine error function'
-      ,'is, conceptually, a compromise between OLS and absolute error'
-      ,'\nfunctions. While at absolute errors of approximately'
-      ,'4 or greater the function is functionally'
-      ,"\nindistinguishable from absolute error, at smaller errors the function's"
-      ,'slope decreases,'
-      ,'\nwhich allows for easier optimization. In practice,'
-      ,'this is a versatile error function that'
[…]
