+++ README.md
@@ -17,6 +17,9 @@ Simple model of exponential change
 A basic model nonlinearly relating time to an outcome variable. The first argument is a data frame, with the first column being the response variable and the second column being the time variable.
 
 ``` r
+# If necessary, install package:
+# devtools::install_github('akcochrane/TEfits')
+
 # generate artificial data:
 dat_simple <- data.frame(response=log(2:31),trial_number=1:30)
 
@@ -86,19 +89,19 @@ summary(mod_boot)
     ## 
     ## >> Fit Values:
     ##        Estimate  Q025  Q975 pseudoSE
-    ## pAsym     0.998 0.985 1.000    0.004
-    ## pRate     2.730 2.544 2.808    0.067
-    ## pStart    0.238 0.144 0.284    0.036
+    ## pAsym     0.999 0.983 1.000    0.004
+    ## pRate     2.707 2.586 2.847    0.067
+    ## pStart    0.224 0.179 0.271    0.023
     ## 
     ## >> Goodness-of-fit:
     ##                err  nullErr nPars nObs      BIC  nullBIC    deltaBIC
-    ## bernoulli 13.42604 16.83409     3   30 37.05566 37.06937 -0.01370674
+    ## bernoulli 13.42474 16.83409     3   30 37.05307 37.06937 -0.01629702
     ## 
     ## >> Test of change in nonindependence:
     ##                          rawSpearman modelConditionalSpearman
-    ## response ~ trial_number:          -1               -0.1497219
+    ## response ~ trial_number:          -1             -0.002447164
     ##                          proportionalSpearmanChange pValSpearmanChange
-    ## response ~ trial_number:                  0.1497219                  0
+    ## response ~ trial_number:                0.002447164                  0
     ##                          pval_KPSS_null pval_KPSS_model
     ## response ~ trial_number:          < .01            > .1
     ## 
@@ -108,10 +111,10 @@ summary(mod_boot)
     ## 
     ## >> Bootstrapped parameter correlations:
     ##         pAsym pStart pRate    err
-    ## pAsym   1.000 -0.305 0.153 -0.160
-    ## pStart -0.305  1.000 0.652  0.564
-    ## pRate   0.153  0.652 1.000  0.381
-    ## err    -0.160  0.564 0.381  1.000
+    ## pAsym   1.000 -0.211 0.200 -0.341
+    ## pStart -0.211  1.000 0.577  0.726
+    ## pRate   0.200  0.577 1.000  0.296
+    ## err    -0.341  0.726 0.296  1.000
 
 Fitting multiple models
 -----------------------
@@ -149,9 +152,9 @@ summary(mod_4group)
     ## >> Formula: response ~ ((pAsym) + ((pStart) - (pAsym)) * 2^((1 - trial_number)/(2^(pRate))))
     ## 
     ## >> Overall effects:
-    ##            pAsym     pStart      pRate
-    ## mean   0.1492270 0.01639027 3.83366264
-    ## stdErr 0.0393339 0.01060448 0.02431836
+    ##             pAsym     pStart      pRate
+    ## mean   0.14922724 0.01639027 3.83366608
+    ## stdErr 0.03933406 0.01060449 0.02431492
     ## 
     ##                 err    nullErr nPars nObs      Fval         Pval   Rsquared
     ## mean   3.005041e-04 0.03071614     3   30 1692.5939 1.110223e-16 0.97598962
@@ -193,7 +196,7 @@ TElm parameter estimates:
 
 |  X.Intercept.|  trial\_number|   rate|
 |-------------:|--------------:|------:|
-|         3.522|         -2.653|  2.867|
+|         3.541|          -2.66|  2.899|
 
 TEfit parameter estimates:
 
