### 2020-08-03 #2535 @paul-buerkner
I just finished my review and have the following remaining points:

- I couldn't find any installation instructions on the README or elsewhere. Did I just overlook them? (Of course I know how to install from github but other people may not).
- Is there any plan to publish this package on a maintained package repository such as CRAN? Of course, this it not required for the JOSS submission but more of a question I have out of interest.
