### 2020-08-03 e2f4fea
update README
include installation instructions
