+++ README.md
@@ -11,14 +11,20 @@ Behavioral data is described, interpreted, and tested using indices such as d pr
 
 The **TEfits** package has a heavy emphasis on interpretability of parameters. As far as possible, parameters fit by **TEfits** are meant to reflect human-interpretable representations of time-evolving processes. Error functions, nonlinear ("change") functions linking predicted values to parameters and time, parameter and prediction boundaries, and goodness-of-fit indices are intended to be clear and adjustable. An equal emphasis is on ease of use: minimal arguments are necessary to begin using the primary function, `TEfit()`, and many common tasks are fully automated (e.g., optimization starting points, bootstrapping).
 
+Installing and loading the package
+----------------------------------
+
+The R package `devtools` includes a very easy way to install packages from Github.
+
+    devtools::install_github('akcochrane/TEfits')
+
 Simple model of exponential change
 ----------------------------------
 
 A basic model nonlinearly relating time to an outcome variable. The first argument is a data frame, with the first column being the response variable and the second column being the time variable.
 
 ``` r
-# If necessary, install package:
-# devtools::install_github('akcochrane/TEfits')
+library(TEfits)
 
 # generate artificial data:
 dat_simple <- data.frame(response=log(2:31),trial_number=1:30)
@@ -29,7 +35,7 @@ mod_simple <- TEfit(dat_simple[,c('response','trial_number')])
 plot(mod_simple,plot_title='Time-evolving fit of artificial data')
 ```
 
-![](README_files/figure-markdown_github/simple_model-1.png)
+![](README_files/figure-markdown_github/model_simple-1.png)
 
 ``` r
 summary(mod_simple)
@@ -89,19 +95,19 @@ summary(mod_boot)
     ## 
     ## >> Fit Values:
     ##        Estimate  Q025  Q975 pseudoSE
-    ## pAsym     0.999 0.983 1.000    0.004
-    ## pRate     2.707 2.586 2.847    0.067
-    ## pStart    0.224 0.179 0.271    0.023
+    ## pAsym     0.997 0.984 1.000    0.004
+    ## pRate     2.712 2.539 2.792    0.065
+    ## pStart    0.236 0.139 0.276    0.035
     ## 
     ## >> Goodness-of-fit:
-    ##                err  nullErr nPars nObs      BIC  nullBIC    deltaBIC
-    ## bernoulli 13.42474 16.83409     3   30 37.05307 37.06937 -0.01629702
+    ##                err  nullErr nPars nObs     BIC  nullBIC    deltaBIC
+    ## bernoulli 13.42761 16.83409     3   30 37.0588 37.06937 -0.01056722
     ## 
     ## >> Test of change in nonindependence:
     ##                          rawSpearman modelConditionalSpearman
-    ## response ~ trial_number:          -1             -0.002447164
+    ## response ~ trial_number:          -1               -0.1697442
     ##                          proportionalSpearmanChange pValSpearmanChange
-    ## response ~ trial_number:                0.002447164                  0
+    ## response ~ trial_number:                  0.1697442                  0
     ##                          pval_KPSS_null pval_KPSS_model
     ## response ~ trial_number:          < .01            > .1
     ## 
@@ -110,11 +116,11 @@ summary(mod_boot)
     ## >> Timepoint at which resampled estimates diverge from timepoint 1, with Cohen's d>1: 2 
     ## 
     ## >> Bootstrapped parameter correlations:
-    ##         pAsym pStart pRate    err
[…]
