+++ README.Rmd
@@ -6,11 +6,6 @@ output:
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-
-```{r, include=FALSE}
-library(TEfits)
-```
-
 # TEfits
 
 [![DOI](https://zenodo.org/badge/225967950.svg)](https://zenodo.org/badge/latestdoi/225967950)
@@ -22,14 +17,21 @@ Behavioral data is described, interpreted, and tested using indices such as d pr
 
 The **TEfits** package has a heavy emphasis on interpretability of parameters. As far as possible, parameters fit by **TEfits** are meant to reflect human-interpretable representations of time-evolving processes. Error functions, nonlinear ("change") functions linking predicted values to parameters and time, parameter and prediction boundaries, and goodness-of-fit indices are intended to be clear and adjustable. An equal emphasis is on ease of use: minimal arguments are necessary to begin using the primary function, `TEfit()`, and many common tasks are fully automated (e.g., optimization starting points, bootstrapping).
 
+## Installing and loading the package
+
+The R package `devtools` includes a very easy way to install packages from Github.
+
+```
+devtools::install_github('akcochrane/TEfits')
+```
+
 ## Simple model of exponential change
 
 A basic model nonlinearly relating time to an outcome variable. The first argument is a data frame, with the first column being the response variable and the second column being the time variable.
 
-```{r simple_model, message=FALSE, warning=FALSE}
+```{r model_simple, message=FALSE, warning=FALSE}
 
-# If necessary, install package:
-# devtools::install_github('akcochrane/TEfits')
+library(TEfits)
 
 # generate artificial data:
 dat_simple <- data.frame(response=log(2:31),trial_number=1:30)
