### 2020-08-03 207cf86
Update README
clarify installation instructions
