### 2020-08-03 #2535 @akcochrane
@paul-buerkner Thank you for the input.

- I've updated the README to include suggested installation instructions.

- I may try to publish the package on CRAN at some point. I would want to re-write much of the package first, however, to make improvements to the interface and performance as well as ensure conformity to CRAN guidelines. I will not be able to manage that workload alone for the foreseeable future.

- Thanks for pointing that out. I'll make a quick pass through the examples today, although it's still likely that I'll miss things occasionally.
