+++ man/plot.TEfit.Rd
@@ -34,6 +34,9 @@ Plot predicted values of a TEfit model. Predictions are thresholds, if relevant;
 otherwise they are overall model predictions
 }
 \examples{
-\dontrun{plot(model_fit_by_TEfit)}
+\dontrun{
+m <- TEfit(anstrain_s1[,c('acc','trialNum')])
+plot(m)
+}
 
 }
