### 2020-08-15 #2535 @ejhigson
Hi @akcochrane - congratulations on a nice software package! I only have a few minor comments:

1. It might be worth adding instructions on running the package's tests to the install section of the README.
2. Some of the references are missing DOIs - please can you add these if they are available (no worries if not). I had thought that commenting "@whedon check references" would check your references automatically but it doesn't seem to have worked just now.
3. I saw a couple of minor typos in the paper:

- last sentence of first paragraph: "allows" -> "allowing".
- first sentence of last paragraph: "as well as having results using TEfits" -> "and results using TEfits have been"?
