### 2020-08-15 a324034
Update test.check_TEfit.R
