### 2020-08-15 #2535 @akcochrane
Thank you for the input @ejhigson ! I've implemented your suggestions.

Just out of curiosity, is it typical to include instructions for tests to be included in high-level introductory documentation? In my minimal experience tests had struck me as being developer-oriented, but knowing more about user-oriented tests would probably help me write better docs and tests!
