### 2020-08-16 #2535 @ejhigson
Thank you for this @akcochrane! Personally I like to include a line for how to run tests in software install documentation (you can always say running tests is optional). If you want to keep the start of the README high-level and concise then I would be happy for you to include the test instructions elsewhere in the documentation instead if you prefer - up to you.

Although the package seems to generally be working running the examples, I get a "no tests found" error when running the tests with test_package (below). Do you know why this is?
```
> testthat::test_package('TEfits')
Error: No tests found for TEfits
```
For reference, here are details of my session:
```
> sessionInfo()
R version 4.0.2 (2020-06-22)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 18363)

Matrix products: default

locale:
[1] LC_COLLATE=English_United Kingdom.1252
[2] LC_CTYPE=English_United Kingdom.1252
[3] LC_MONETARY=English_United Kingdom.1252
[4] LC_NUMERIC=C
[5] LC_TIME=English_United Kingdom.1252

attached base packages:

other attached packages:
[1] TEfits_00.77.07 testthat_2.3.2

[1] compiler_4.0.2 magrittr_1.5   R6_2.4.1       rlang_0.4.7
```
