+++ tests/testthat/test.check_TEfitAll.R
@@ -8,51 +8,98 @@ d <- anstrain
 d$absRat <- abs(d$ratio)
 d$moreCatA <- as.numeric(d$ratio > 0)
 
-m <- list()
-test_that('TEfitAll runs with various link and error functions',{
+### ###
+test_that('TEfitAll runs with identity link and OLS error function',{
+  expect_is({
+    m$ident_ols <- TEfitAll(d[,c('acc','trialNum')],
+                            bootPars = tef_bootList(resamples = 20),groupingVar = d$subID)
+    m$ident_ols},
+    'TEfitAll')
+})
 
-  expect_success({
-m$ident_ols <- TEfitAll(d[,c('acc','trialNum')],
-                        bootPars = tef_bootList(resamples = 20),groupingVar = d$subID)
-  })
-  expect_success({
+
+### ###
+test_that('TEfitAll runs with identity link and logcosh error function',{
+  expect_is({
     m$ident_logcosh <- TEfitAll(d[,c('acc','trialNum')],
                                 bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
-                             errFun = 'logcosh')
-  })
-  expect_success({
-m$ident_bern <- TEfitAll(d[,c('acc','trialNum')],
+                                errFun = 'logcosh')
+    m$ident_logcosh},
+    'TEfitAll')
+})
+
+
+### ###
+test_that('TEfitAll runs with identity link and bernoulli error function',{
+
+  expect_is({
+    m$ident_bern <- TEfitAll(d[,c('acc','trialNum')],
+                             bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
+                             errFun = 'bernoulli')
+    m$ident_bern},
+    'TEfitAll')
+})
+
+
+### ###
+test_that('TEfitAll runs with Weibull link and OLS error function',{
+
+  expect_is({
+    m$weib_ols <- TEfitAll(d[,c('acc','trialNum','absRat')],
+                           bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
+                           linkFun = list(link='weibull',weibullX = 'absRat')
+    )
+    m$weib_ols},
+    'TEfitAll')
+})
+
+### ###
+test_that('TEfitAll runs with Weibull link and bernoulli error function',{
+
+  expect_is({
+    m$weib_bern <- TEfitAll(d[,c('acc','trialNum','absRat')],
+                            bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
+                            linkFun = list(link='weibull',weibullX = 'absRat'),
+                            errFun = 'bernoulli')
+    m$weib_bern},
+    'TEfitAll')
+})
+
+### ###
+test_that('TEfitAll runs with logistic link and OLS error function',{
+  expect_is({
+    m$logist_ols <- TEfitAll(d[,c('acc','trialNum','ratio')],
+                             bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
+                             linkFun = list(link='logit',logistX = 'ratio')
+    )
+    m$logist_ols},
+    'TEfitAll')
+})
+
+### ###
+test_that('TEfitAll runs with logistic link and bernoulli error function',{
+
+  expect_is({
+    m$logist_bern <- TEfitAll(d[,c('acc','trialNum','ratio')],
+                              bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
+                              linkFun = list(link='logit',logistX = 'absRat'),
+                              errFun = 'bernoulli')
+    m$logist_bern},
+    'TEfitAll')
+})
+
+### ###
+test_that('TEfitAll runs with d prime link function',{
+
+  expect_is({
+    m$dPrime <- TEfitAll(d[,c('acc','trialNum','moreCatA')],
[…]
