+++ DESCRIPTION
@@ -9,6 +9,11 @@ License: MIT
 Encoding: UTF-8
 LazyData: true
 Imports: graphics, MASS, stats
-Suggests: knitr, rmarkdown, psych, lme4
+Suggests: 
+    knitr,
+    rmarkdown,
+    psych,
+    lme4,
+    testthat
 VignetteBuilder: knitr
 RoxygenNote: 7.1.0
