### 2020-08-16 4194c62
Update test.check_TEfitAll.R
