+++ tests/testthat/test.check_TEfitAll.R
@@ -3,6 +3,7 @@
 context("check_TEfitAll")
 
 library(TEfits)
+
 d <- anstrain
 
 d$absRat <- abs(d$ratio)
