+++ tests/testthat/test.check_TEfitAll.R
@@ -9,11 +9,14 @@ d <- anstrain
 d$absRat <- abs(d$ratio)
 d$moreCatA <- as.numeric(d$ratio > 0)
 
+m <- list()
+
 ### ###
 test_that('TEfitAll runs with identity link and OLS error function',{
   expect_is({
     m$ident_ols <- TEfitAll(d[,c('acc','trialNum')],
-                            bootPars = tef_bootList(resamples = 20),groupingVar = d$subID)
+                            bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
+                            tef_control(nTries = 50))
     m$ident_ols},
     'TEfitAll')
 })
@@ -21,10 +24,12 @@ test_that('TEfitAll runs with identity link and OLS error function',{
 
 ### ###
 test_that('TEfitAll runs with identity link and logcosh error function',{
+  skip_on_travis()
   expect_is({
     m$ident_logcosh <- TEfitAll(d[,c('acc','trialNum')],
                                 bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
-                                errFun = 'logcosh')
+                                errFun = 'logcosh',
+                                tef_control(nTries = 50))
     m$ident_logcosh},
     'TEfitAll')
 })
@@ -32,11 +37,12 @@ test_that('TEfitAll runs with identity link and logcosh error function',{
 
 ### ###
 test_that('TEfitAll runs with identity link and bernoulli error function',{
-
+  skip_on_travis()
   expect_is({
     m$ident_bern <- TEfitAll(d[,c('acc','trialNum')],
                              bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
-                             errFun = 'bernoulli')
+                             errFun = 'bernoulli',
+                             tef_control(nTries = 50))
     m$ident_bern},
     'TEfitAll')
 })
@@ -44,11 +50,12 @@ test_that('TEfitAll runs with identity link and bernoulli error function',{
 
 ### ###
 test_that('TEfitAll runs with Weibull link and OLS error function',{
-
+  skip_on_travis()
   expect_is({
     m$weib_ols <- TEfitAll(d[,c('acc','trialNum','absRat')],
                            bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
-                           linkFun = list(link='weibull',weibullX = 'absRat')
+                           linkFun = list(link='weibull',weibullX = 'absRat'),
+                           tef_control(nTries = 50)
     )
     m$weib_ols},
     'TEfitAll')
@@ -61,17 +68,20 @@ test_that('TEfitAll runs with Weibull link and bernoulli error function',{
     m$weib_bern <- TEfitAll(d[,c('acc','trialNum','absRat')],
                             bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                             linkFun = list(link='weibull',weibullX = 'absRat'),
-                            errFun = 'bernoulli')
+                            errFun = 'bernoulli',
+                            tef_control(nTries = 50))
     m$weib_bern},
     'TEfitAll')
 })
 
 ### ###
 test_that('TEfitAll runs with logistic link and OLS error function',{
+  skip_on_travis()
   expect_is({
     m$logist_ols <- TEfitAll(d[,c('acc','trialNum','ratio')],
                              bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
-                             linkFun = list(link='logit',logistX = 'ratio')
+                             linkFun = list(link='logit',logistX = 'ratio'),
+                             tef_control(nTries = 50)
     )
     m$logist_ols},
     'TEfitAll')
@@ -84,18 +94,19 @@ test_that('TEfitAll runs with logistic link and bernoulli error function',{
[…]
