### 2020-08-16 66de58c
Update test.check_TEfitAll.R
