### 2020-08-16 #2535 @akcochrane
Thanks for pointing that out. I'm not super familiar with the intricacies of devtools, but it looks like installing locally (i.e., `install('dir')` includes tests but installing from github (i.e., `install_github('repo')` can't. In that case, it looks like users might need to download the repo and use `devtools::install()` from there.

I'd be happy to hear other suggestions, particularly if there's an option for `install_github()` to include tests. I don't see anything in the `install_github()` docs that refers to tests, though.

If the above seems like a good way to suggest tests to users then I'll write up some instructions saying, in effect, that users will need to download/clone the repo and install locally.
