+++ tests/testthat/test.check_TEfitAll.R
@@ -24,7 +24,7 @@ test_that('TEfitAll runs with identity link and OLS error function',{
 
 ### ###
 test_that('TEfitAll runs with identity link and logcosh error function',{
-  skip_on_travis()
+  skip_on_travis() # times out
   expect_is({
     m$ident_logcosh <- TEfitAll(d[,c('acc','trialNum')],
                                 bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
@@ -37,7 +37,7 @@ test_that('TEfitAll runs with identity link and logcosh error function',{
 
 ### ###
 test_that('TEfitAll runs with identity link and bernoulli error function',{
-  skip_on_travis()
+  skip_on_travis() # times out
   expect_is({
     m$ident_bern <- TEfitAll(d[,c('acc','trialNum')],
                              bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
@@ -50,7 +50,7 @@ test_that('TEfitAll runs with identity link and bernoulli error function',{
 
 ### ###
 test_that('TEfitAll runs with Weibull link and OLS error function',{
-  skip_on_travis()
+  skip_on_travis() # times out
   expect_is({
     m$weib_ols <- TEfitAll(d[,c('acc','trialNum','absRat')],
                            bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
@@ -63,7 +63,7 @@ test_that('TEfitAll runs with Weibull link and OLS error function',{
 
 ### ###
 test_that('TEfitAll runs with Weibull link and bernoulli error function',{
-  skip_on_travis()
+  skip_on_travis() # times out
   expect_is({
     m$weib_bern <- TEfitAll(d[,c('acc','trialNum','absRat')],
                             bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
@@ -76,7 +76,7 @@ test_that('TEfitAll runs with Weibull link and bernoulli error function',{
 
 ### ###
 test_that('TEfitAll runs with logistic link and OLS error function',{
-  skip_on_travis()
+  skip_on_travis() # times out
   expect_is({
     m$logist_ols <- TEfitAll(d[,c('acc','trialNum','ratio')],
                              bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
@@ -89,11 +89,11 @@ test_that('TEfitAll runs with logistic link and OLS error function',{
 
 ### ###
 test_that('TEfitAll runs with logistic link and bernoulli error function',{
-  skip_on_travis()
+  skip_on_travis() # times out
   expect_is({
     m$logist_bern <- TEfitAll(d[,c('acc','trialNum','ratio')],
                               bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
-                              linkFun = list(link='logit',logistX = 'absRat'),
+                              linkFun = list(link='logit',logistX = 'ratio'),
                               errFun = 'bernoulli',
                               tef_control(nTries = 50))
     m$logist_bern},
@@ -102,14 +102,14 @@ test_that('TEfitAll runs with logistic link and bernoulli error function',{
 
 ### ###
 test_that('TEfitAll runs with d prime link function',{
-  skip_on_travis()
+  skip_on_travis() # times out
   expect_is({
     m$dPrime <- TEfitAll(d[,c('acc','trialNum','moreCatA')],
                          bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                          linkFun = list(link='d_prime',presence = 'moreCatA'),
                          tef_control(nTries = 50)
     )
-    m$drime},
+    m$dPrime},
     'TEfitAll')
 })
 
