### 2020-08-16 29a799d
update README and tests
