+++ R/tef_tryFits.R
@@ -41,34 +41,17 @@ tef_tryFits <- function(modList,whichPnames='pNames',whichFun='evalFun'){
 
 
     guesses[grep('sigma_param',modList$guessNames)] <- runif(length(grep('sigma_param',modList$guessNames)),
-                                                     0,mad(modList$varIn[,modList$respVar],na.rm=T))
+                                                             0,mad(modList$varIn[,modList$respVar],na.rm=T))
 
     guesses[grep('mu_param',modList$guessNames)] <- runif(length(grep('mu_param',modList$guessNames)),
-                                                  0,min(modList$varIn[,modList$respVar],na.rm=T))
+                                                          0,min(modList$varIn[,modList$respVar],na.rm=T))
     guesses[grep('tau_param',modList$guessNames)] <- runif(length(grep('tau_param',modList$guessNames)),
-                                                   0,min(modList$varIn[,modList$respVar],na.rm=T))
+                                                           0,min(modList$varIn[,modList$respVar],na.rm=T))
 
     names(guesses) <- modList$guessNames
     {
 
-        tryCatch({
-           curFit <- optim(guesses,fn=tef_fitErr,
-                          varIn=modList$varIn,pNames=modList$guessNames,evalFun=modList[[whichFun]],
-                          errFun=modList$errFun,respVar=modList$respVar,linkFunX=linkFunX,
-                          y_lim=modList$y_lim,rate_lim=modList$rate_lim,
-                          shape_lim=modList$shape_lim,
-                          penalizeMean = c(penalizeMean,mean(nullYhat,na.rm=T)),
-                          penalizeRate = modList$penalizeRate,
-                          parLims=modList$parLims,
-                          thresh_fun = thresh_fun,
-                          paramTerms = paramTerms,
-                          upper = modList$parLims$parMax,
-                          lower = modList$parLims$parMin,
-                          method='L-BFGS-B',
-                          control=list(maxit=100)
-           )
-        },error = function(.){try({
-          if(modList$quietErrs){cat('\nL-BFGS-B failed')}
+      tryCatch({
         curFit <- optim(guesses,fn=tef_fitErr,
                         varIn=modList$varIn,pNames=modList$guessNames,evalFun=modList[[whichFun]],
                         errFun=modList$errFun,respVar=modList$respVar,linkFunX=linkFunX,
@@ -79,11 +62,28 @@ tef_tryFits <- function(modList,whichPnames='pNames',whichFun='evalFun'){
                         parLims=modList$parLims,
                         thresh_fun = thresh_fun,
                         paramTerms = paramTerms,
-                         method='BFGS',  # use this or L-BFGS-B (with upper and lower) if bounds have been figured out.
+                        upper = modList$parLims$parMax,
+                        lower = modList$parLims$parMin,
+                        method='L-BFGS-B',
                         control=list(maxit=100)
         )
-        },silent=T)
-        })
+      },error = function(.){try({
+        if(modList$quietErrs){cat('\nL-BFGS-B failed')}
+        curFit <- optim(guesses,fn=tef_fitErr,
+                        varIn=modList$varIn,pNames=modList$guessNames,evalFun=modList[[whichFun]],
[…]
