+++ tests/testthat/test.check_TEfitAll.R
@@ -16,7 +16,7 @@ test_that('TEfitAll runs with identity link and OLS error function',{
   expect_is({
     m$ident_ols <- TEfitAll(d[,c('acc','trialNum')],
                             bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
-                            tef_control(nTries = 50))
+                            control = tef_control(suppressWarnings = T,nTries = 50))
     m$ident_ols},
     'TEfitAll')
 })
@@ -29,7 +29,7 @@ test_that('TEfitAll runs with identity link and logcosh error function',{
     m$ident_logcosh <- TEfitAll(d[,c('acc','trialNum')],
                                 bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                                 errFun = 'logcosh',
-                                tef_control(nTries = 50))
+                                control = tef_control(suppressWarnings = T,nTries = 50))
     m$ident_logcosh},
     'TEfitAll')
 })
@@ -42,7 +42,7 @@ test_that('TEfitAll runs with identity link and bernoulli error function',{
     m$ident_bern <- TEfitAll(d[,c('acc','trialNum')],
                              bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                              errFun = 'bernoulli',
-                             tef_control(nTries = 50))
+                             control = tef_control(suppressWarnings = T,nTries = 50))
     m$ident_bern},
     'TEfitAll')
 })
@@ -55,7 +55,7 @@ test_that('TEfitAll runs with Weibull link and OLS error function',{
     m$weib_ols <- TEfitAll(d[,c('acc','trialNum','absRat')],
                            bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                            linkFun = list(link='weibull',weibullX = 'absRat'),
-                           tef_control(nTries = 50)
+                           control = tef_control(suppressWarnings = T,nTries = 50)
     )
     m$weib_ols},
     'TEfitAll')
@@ -69,7 +69,7 @@ test_that('TEfitAll runs with Weibull link and bernoulli error function',{
                             bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                             linkFun = list(link='weibull',weibullX = 'absRat'),
                             errFun = 'bernoulli',
-                            tef_control(nTries = 50))
+                            control = tef_control(suppressWarnings = T,nTries = 50))
     m$weib_bern},
     'TEfitAll')
 })
@@ -81,7 +81,7 @@ test_that('TEfitAll runs with logistic link and OLS error function',{
     m$logist_ols <- TEfitAll(d[,c('acc','trialNum','ratio')],
                              bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                              linkFun = list(link='logit',logistX = 'ratio'),
-                             tef_control(nTries = 50)
+                             control = tef_control(suppressWarnings = T,nTries = 50)
     )
     m$logist_ols},
     'TEfitAll')
@@ -95,7 +95,7 @@ test_that('TEfitAll runs with logistic link and bernoulli error function',{
                               bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
                               linkFun = list(link='logit',logistX = 'ratio'),
                               errFun = 'bernoulli',
-                              tef_control(nTries = 50))
+                              control = tef_control(suppressWarnings = T,nTries = 50))
     m$logist_bern},
     'TEfitAll')
 })
@@ -107,7 +107,7 @@ test_that('TEfitAll runs with d prime link function',{
     m$dPrime <- TEfitAll(d[,c('acc','trialNum','moreCatA')],
                          bootPars = tef_bootList(resamples = 20),groupingVar = d$subID,
[…]
