+++ README.Rmd
@@ -27,7 +27,7 @@ The R package `devtools` includes a very easy way to install packages from Githu
 devtools::install_github('akcochrane/TEfits', build_vignettes = TRUE)
 ```
 
-Although having vignettes is nice for exploring the functionality of the package, building the vignettes takes a minute or two. Remove the `build_vignettes = TRUE` argument to speed up installation.
+Although having vignettes is nice for exploring the functionality of the package (via `browseVignettes('TEfits')`), building the vignettes takes a minute or two. Remove the `build_vignettes = TRUE` argument to speed up installation.
 
 ## Simple model of exponential change
 
@@ -117,7 +117,7 @@ start as an offset from asymptote (ie., `Intercept`).
 
 # Testing functionality
 
-`TEfits` includes automatic testing using the `testthat` package and [Travis-CI](https://travis-ci.com/github/akcochrane/TEfits). If users wish to run these tests locall, it's recommended to download/clone the repo to a local `~/TEfits`. Then install and run tests as follows:
+`TEfits` includes automatic testing using the `testthat` package and [Travis-CI](https://travis-ci.com/github/akcochrane/TEfits). If users wish to run these tests locally, it's recommended to download/clone the repo to a local directory `~/TEfits`. Then install and run tests as follows:
 
 ```
 devtools::install('~/TEfits') # replace '~' with your filepath
