+++ README.md
@@ -18,7 +18,7 @@ The R package `devtools` includes a very easy way to install packages from Githu
 
     devtools::install_github('akcochrane/TEfits', build_vignettes = TRUE)
 
-Although having vignettes is nice for exploring the functionality of the package, building the vignettes takes a minute or two. Remove the `build_vignettes = TRUE` argument to speed up installation.
+Although having vignettes is nice for exploring the functionality of the package (via `browseVignettes('TEfits')`), building the vignettes takes a minute or two. Remove the `build_vignettes = TRUE` argument to speed up installation.
 
 Simple model of exponential change
 ----------------------------------
@@ -81,6 +81,12 @@ dat <- data.frame(response=log(2:31)/log(32),trial_number=1:30)
 mod_boot <- TEfit(dat[,c('response','trial_number')], 
              errFun='bernoulli',
              bootPars=tef_bootList(resamples = 40))
+```
+
+    ## 
+    ## Warning: model did not converge at tol = 0.05 . Consider respecifying, allowing more runs, or increasing the convergence tolerance.
+
+``` r
 plot(mod_boot,plot_title='Time-evolving fit of artificial data with 95% CI from 40 bootstrapped fits')
 ```
 
@@ -93,23 +99,24 @@ summary(mod_boot)
     ## 
     ## >> Formula: response~((pAsym) + ((pStart) - (pAsym)) * 2^((1 - trial_number)/(2^(pRate))))
     ## 
-    ## >> Converged: TRUE 
+    ## >> Converged: FALSE 
+    ## >> Max runs: 200  -- Tolerance: 0.05 
     ## 
     ## >> Fit Values:
     ##        Estimate  Q025  Q975 pseudoSE
-    ## pAsym     0.996 0.987 1.000    0.003
-    ## pRate     2.721 2.610 2.816    0.053
-    ## pStart    0.229 0.201 0.278    0.020
+    ## pAsym     0.996 0.984 1.000    0.004
+    ## pRate     2.737 2.594 2.765    0.044
+    ## pStart    0.231 0.123 0.271    0.038
     ## 
     ## >> Goodness-of-fit:
-    ##                err  nullErr nPars nObs      BIC  nullBIC    deltaBIC
-    ## bernoulli 13.42677 16.83409     3   30 37.05713 37.06937 -0.01224387
+    ##                err  nullErr nPars nObs      BIC  nullBIC     deltaBIC
+    ## bernoulli 13.42851 16.83409     3   30 37.06062 37.06937 -0.008755303
     ## 
     ## >> Test of change in nonindependence:
     ##                          rawSpearman modelConditionalSpearman
-    ## response ~ trial_number:          -1              -0.07096774
+    ## response ~ trial_number:          -1              -0.09410456
     ##                          proportionalSpearmanChange pValSpearmanChange
-    ## response ~ trial_number:                 0.07096774                  0
+    ## response ~ trial_number:                 0.09410456                  0
     ##                          pval_KPSS_null pval_KPSS_model
     ## response ~ trial_number:          < .01            > .1
     ## 
@@ -118,11 +125,11 @@ summary(mod_boot)
     ## >> Timepoint at which resampled estimates diverge from timepoint 1, with Cohen's d>1: 2 
     ## 
     ## >> Bootstrapped parameter correlations:
-    ##         pAsym pStart  pRate    err
-    ## pAsym   1.000 -0.300  0.097 -0.426
-    ## pStart -0.300  1.000  0.685  0.291
-    ## pRate   0.097  0.685  1.000 -0.138
-    ## err    -0.426  0.291 -0.138  1.000
+    ##        pAsym pStart pRate   err
[…]
