### 2020-08-16 #2535 @ejhigson
Installing locally with ```devtools::install()``` from a download of the latest version of master from github makes the test run for me! Please can you note this in the documentation somewhere to avoid confusion for users who want to run the tests? I get some errors when I run the tests though due to "undefined columns selected" - do you get these too/do you know why they are occuring?
```
> testthat::test_package('TEfits')

Your rate is very close to the boundary. Consider penalizing the likelihood.-- 1. Error: TEfitAll runs with identity link and OLS error function (@test.chec
undefined columns selected
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:16:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:17:4
 8. base::`[.data.frame`(...)

Your rate is very close to the boundary. Consider penalizing the likelihood.-- 2. Error: TEfitAll runs with identity link and logcosh error function (@test.
undefined columns selected
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:28:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:29:4
 8. base::`[.data.frame`(...)

Your rate is very close to the boundary. Consider penalizing the likelihood.-- 3. Error: TEfitAll runs with identity link and bernoulli error function (@tes
undefined columns selected
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:41:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:42:4
 8. base::`[.data.frame`(...)

-- 4. Error: TEfitAll runs with Weibull link and OLS error function (@test.check
undefined columns selected
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:54:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:55:4
 8. base::`[.data.frame`(...)

-- 5. Error: TEfitAll runs with Weibull link and bernoulli error function (@test
undefined columns selected
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:67:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:68:4
 8. base::`[.data.frame`(...)

Warning: model did not converge at tol = 0.05 . Consider respecifying, allowing more runs, or increasing the convergence tolerance.
-- 6. Error: TEfitAll runs with logistic link and OLS error function (@test.chec
undefined columns selected
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:80:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:81:4
 8. base::`[.data.frame`(...)

Warning: model did not converge at tol = 0.05 . Consider respecifying, allowing more runs, or increasing the convergence tolerance.
-- 7. Error: TEfitAll runs with logistic link and bernoulli error function (@tes
object 'absRat' not found
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:93:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:94:4
 5. TEfits::TEfit(...)
 6. TEfits::tef_tryFits(modList, whichPnames = "null_pNames", whichFun = "null_fun")
 7. TEfits::tef_fitErr(...)
 8. [ base::eval(...) ] with 1 more call

Your rate is very close to the boundary. Consider penalizing the likelihood.-- 8. Error: TEfitAll runs with d prime link function (@test.check_TEfitAll.R#10
undefined columns selected
Backtrace:
 1. testthat::expect_is(...) tests/testthat/test.check_TEfitAll.R:106:2
 4. TEfits::TEfitAll(...) tests/testthat/test.check_TEfitAll.R:107:4
 8. base::`[.data.frame`(...)

== testthat results  ===========================================================
[ OK: 45 | SKIPPED: 0 | WARNINGS: 4 | FAILED: 8 ]
1. Error: TEfitAll runs with identity link and OLS error function (@test.check_TEfitAll.R#16)
2. Error: TEfitAll runs with identity link and logcosh error function (@test.check_TEfitAll.R#28)
3. Error: TEfitAll runs with identity link and bernoulli error function (@test.check_TEfitAll.R#41)
4. Error: TEfitAll runs with Weibull link and OLS error function (@test.check_TEfitAll.R#54)
5. Error: TEfitAll runs with Weibull link and bernoulli error function (@test.check_TEfitAll.R#67)
6. Error: TEfitAll runs with logistic link and OLS error function (@test.check_TEfitAll.R#80)
7. Error: TEfitAll runs with logistic link and bernoulli error function (@test.check_TEfitAll.R#93)
8. Error: TEfitAll runs with d prime link function (@test.check_TEfitAll.R#106)

Error: testthat unit tests failed
```
Here is my session info:
```
> sessionInfo()
R version 4.0.2 (2020-06-22)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 18363)

Matrix products: default

locale:
[1] LC_COLLATE=English_United Kingdom.1252
[2] LC_CTYPE=English_United Kingdom.1252
[3] LC_MONETARY=English_United Kingdom.1252
[4] LC_NUMERIC=C
[5] LC_TIME=English_United Kingdom.1252

attached base packages:

other attached packages:
[1] TEfits_00.77.07 testthat_2.3.2

 [1] ps_1.3.4          fansi_0.4.1       prettyunits_1.1.1 rprojroot_1.3-2
 [5] withr_2.2.0       digest_0.6.25     crayon_1.3.4      assertthat_0.2.1
 [9] R6_2.4.1          backports_1.1.7   magrittr_1.5      rlang_0.4.7
[13] cli_2.0.2         fs_1.5.0          remotes_2.2.0     callr_3.4.3
[17] ellipsis_0.3.1    desc_1.2.0        devtools_2.3.1    tools_4.0.2
[21] glue_1.4.1        pkgload_1.1.0     compiler_4.0.2    processx_3.4.3
[25] pkgbuild_1.1.0    sessioninfo_1.1.1 memoise_1.1.0     usethis_1.6.1
```
