### 2020-08-16 #2535 @akcochrane
I included the suggested instructions (to download and run locally) near the end of the README.

Would you mind trying again with the latest version (TEfits_00.77.09)? Thanks!
