### 2020-08-16 #2535 @akcochrane
It appears as though the issue is a R 4.x incompatibility. I'll let you know when that's fixed.
