### 2020-08-16 70bf1fc
allow for TEfitAll without the psych package
