### 2020-08-16 1f252ca
Update TEfitAll.R
