### 2020-08-16 1ba2cea
allow flexibility in TEfitAll dependencies
