+++ DESCRIPTION
@@ -16,4 +16,4 @@ Suggests:
     lme4,
     testthat
 VignetteBuilder: knitr
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
