### 2020-08-16 fc88262
Update test.check_TElm.R
