### 2020-08-16 #2535 @akcochrane
Evidently I'd accidentally left a necessary dependency on a package that I'd meant to be optional. That bug should be fixed now, and the tests should run.
