### 2020-08-17 #2535 @ejhigson
@akcochrane thank you very much for fixing this! The tests now all pass for me.

I am happy to recommend this paper for publication in JOSS. Congratulations on a nice software package!
