### 2020-08-19 #2535 @cMadan
@akcochrane,  looks like that ref should have DOI of 10.1167/jov.20.8.16 and no longer "in press" (https://pubmed.ncbi.nlm.nih.gov/32790849/).

After that, there are a few last things to take care of:

You may find this helpful: https://guides.github.com/activities/citable-code/
