+++ doc/references.bib
@@ -61,8 +61,9 @@
 	title = {Perceptual Learning of Appendicitis Diagnosis in Radiological Images},
 	journal = {Journal of Vision},
 	year = {2020},
-	volume = {in press},
-        doi = {10.1167/jov.0.0.07387},
+	volume = {20},
+	issue = {8},
+        doi = {10.1167/jov.20.8.16},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
 
