+++ doc/references.bib
@@ -62,7 +62,7 @@
 	journal = {Journal of Vision},
 	year = {2020},
 	volume = {20},
-	issue = {8},
+	number = {8},
         doi = {10.1167/jov.20.8.16},
 	author = {Johnston, Ian and Ji, Mohan and Cochrane, Aaron and Demko, Zachary and Robbins, Jessica and Stephenson, Jason and Green, C. Shawn}
 }
