### 2020-08-21 #2535 @akcochrane
No, it was not a typo.
