### 2020-10-04 8c92b00
Improvements to various functions
