+++ R/TElmem.R
@@ -34,6 +34,7 @@
 #' @param timeVar String. Indicates which variable in \code{datIn} corresponds to time (i.e., should be transformed). Must be numeric and positive.
 #' @param groupingVar String. Indicates which variable in \code{datIn} should have a time=related random effect.
 #' @param nRuns Number of times to run optimization of the rate (i.e., fitting nonlinear transformations of \code{timeVar})
+#' @param startingOffset By default (if T) time is coded to start at 1 and saturate to 0. If startingOffset is F, time starts at 0 and saturates to 1. May assist in interpreting interactions with other variables, etc.
 #' @param silent Progress is printed by default. silent=T to suppress
 #'
 #' @return
@@ -55,7 +56,7 @@
 #' # Participant-level rate parameters:
 #' m_TElmem$rates
 #' }
-TElmem <- function(formIn,dat,timeVar,groupingVar,nRuns = 1,silent=F){
+TElmem <- function(formIn,dat,timeVar,groupingVar,nRuns = 1,startingOffset=T,silent=F){
   require(lme4)
 
   minTime <- min(dat[,timeVar],na.rm = T)
@@ -79,6 +80,7 @@ TElmem <- function(formIn,dat,timeVar,groupingVar,nRuns = 1,silent=F){
       groupMods[[curGroupName]] <- TElm(groupForm,
                                         timeVar = timeVar,
                                         dat = dat[dat[,groupingVar]==curGroupName,],
+                                        startingOffset = startingOffset,
                                         nBoot = 20
       )
       rateVect[curGroupName] <- groupMods[[curGroupName]]$rate
@@ -99,6 +101,7 @@ TElmem <- function(formIn,dat,timeVar,groupingVar,nRuns = 1,silent=F){
     for(curGV in groupNames){
       curDat[curDat[,groupingVar]==curGV,timeVar] <-
         2^((minTime-curDat[curDat[,groupingVar]==curGV,timeVar])/(2^bestRates[curGV]))
+      if(!startingOffset){curDat[curDat[,groupingVar]==curGV,timeVar] <- 1-curDat[curDat[,groupingVar]==curGV,timeVar]}
     }
     initMod <- lmer(formIn,curDat)
     bestNegLL <- -logLik(initMod)
@@ -109,10 +112,11 @@ TElmem <- function(formIn,dat,timeVar,groupingVar,nRuns = 1,silent=F){
   if(nRuns==0){cat(']\n') ; return(list(models=groupMods,rates=rateVect,timeDat=timeDat,lmerMod = initMod))}
 
   ## define function to optimize (input a vector of rates to minimize negative L)
-  fitFun <- function(rates,curDat,formula,timeVar,groupNames,rateBounds){
+  fitFun <- function(rates,curDat,formula,timeVar,groupNames,rateBounds,startingOffset){
     for(curGV in groupNames){
       curDat[curDat[,groupingVar]==curGV,timeVar] <-
         2^((minTime-curDat[curDat[,groupingVar]==curGV,timeVar])/(2^rates[curGV]))
+      if(!startingOffset){curDat[curDat[,groupingVar]==curGV,timeVar] <- 1-curDat[curDat[,groupingVar]==curGV,timeVar]}
     }
     modNegLL <- 1E15
     try({
@@ -139,7 +143,8 @@ TElmem <- function(formIn,dat,timeVar,groupingVar,nRuns = 1,silent=F){
                 formula=formIn,
                 timeVar=timeVar,
                 groupNames=groupNames,
-                rateBounds=rateBounds)
+                rateBounds=rateBounds,
[…]
