### 2020-11-20 014738b
Allow fixed effects in models re-fit by brms
