+++ R/tef_fitAll2brms.R
@@ -7,12 +7,16 @@
 #' this function is less tested than most functions in the \code{TEfits} package. Functionality is
 #' \strong{experimental}.
 #'
+#' Priors for nonlinear parameters are informed by the distributions of parameters in the \code{TEfitAll} object [models].
+#' However, any fixed effects should be minimally influenced by these priors
+#'
 #' \code{TEfitAll} \code{bernoulli} models are fit using either \code{bernoulli} or \code{Beta} response
-#' distributions in \code{brms} depending on whether the \code{TEfitAll} distribution is
+#' distributions in \code{brms} depending on whether the \code{TEfitAll} distrIibution is
 #' binary. \code{TEfitAll} \code{logcosh} models are fit using a \code{asym_laplace} response distribution
 #' in brms predicting the .5 quantile.
 #'
 #' @param TEs3s TEfitAll model
+#' @param fixef Parameters vary as random effects by the TEs3s grouping variable. However, if you have main effects (e.g., group differences), enter them \emph{as a data frame} here.
 #' @param nIter number of iterations
 #' @param nChains number of chains
 #' @param nCores number of cores
@@ -24,8 +28,32 @@
 #'
 #' @return A \code{\link[brms]{brms-package}} nonlinear mixed-effects model object.
 #'
+#' @examples
+#' \dontrun{
+#' dat <- anstrain
+#' dat$condition <- rep(c('A','B'),each=500)
+#'
+#' # Model with time and one categorical fixed effect
+#' mod_tef <- TEfitAll(dat[,c('acc','trialNum')], groupingVar = dat$subID)
+#' mod_brm <- tef_fitAll2brms(mod_tef,nChains=1,fixef=data.frame(condition=dat$condition))
+#'
+#' # Model with time, one categorical fixed effect, and one by-groupingVar (subID) random slope
+#' dat$absRat <- scale(abs(dat$ratio))
+#' mod_tef <- TEfitAll(dat[,c('acc','trialNum',"absRat")], groupingVar = dat$subID,covarTerms=list(pRate=c(F)))
+#' mod_brm <- tef_fitAll2brms(mod_tef,nChains=1,fixef=data.frame(condition=dat$condition))
+#' }
+#'
 #' @export
-tef_fitAll2brms <- function(TEs3s,nIter= 2000,nChains=3,nCores=2,errFun=NA,prior_dispersion=2){
+tef_fitAll2brms <- function(TEs3s,fixef=NA,nIter= 2000,nChains=3,nCores=2,errFun=NA,prior_dispersion=2){
+
+  ## To do:
+  ##
+  ## ## have a fixed=c() argument, to wedge between 'pAsym ~' and '(1||groupingVar)' and similar constructions
+  ##
+  ## ## fit in a loop that keeps trying until a fit with samples, so that the user won't get something out with no samples
+  ##
+  ## ## probably should improve the docs, expecially around cores and chains. Someone who's never used brms should be able to get the broad strokes here.
+  ## ## Also, seriously think about defaulting to one core and one chain. Given the likelihood of pathological sampling (the need for looping to sample), it's better.
 
   require(brms)
   # par_lims <- TEs3s$allFitList[[1]]$modList$parLims
@@ -34,17 +62,31 @@ tef_fitAll2brms <- function(TEs3s,nIter= 2000,nChains=3,nCores=2,errFun=NA,prior
 
   groupingVarName <- attr(TEs3s$fitSummary,'grouping_var')
 
+
+
   varIn <- data.frame(); for(curGroup in 1:length(TEs3s$allFitList)){
     subDat <- data.frame(TEs3s$allFitList[[curGroup]]$data)
[…]
