+++ R/TEfit.R
@@ -364,7 +364,12 @@ TEfit <- function(varIn,
     bestFit$GoF$BIC <- log(bestFit$value/nObs)*nObs + bestFit$GoF$nPars*log(nObs)
     bestFit$GoF$nullBIC <- log(nullFit$value/nObs)*nObs + length(nullFit$par)*log(nObs)
     bestFit$GoF$deltaBIC <- bestFit$GoF$BIC - bestFit$GoF$nullBIC
+
+    ## >> next step is BIC-BF approximation: log(exp(deltaBIC/2),base=desired_base)
   }
+
+  ## Need to switch() this
+
   if(modList$errFun == 'rmse'){
     bestFit$GoF$BIC <- log(bestFit$value^2)*nObs + bestFit$GoF$nPars*log(nObs)
     bestFit$GoF$nullBIC <- log(nullFit$value^2)*nObs + length(nullFit$par)*log(nObs)
